move Reports to suspense
