+++ package.json
@@ -30,7 +30,6 @@
         "qs": "^6.5.2",
         "react": "^16.6.0",
         "react-dom": "^16.6.0",
-        "react-loadable": "^5.5.0",
         "react-redux": "^5.0.6",
         "react-router": "^4.3.1",
         "react-router-dom": "^4.3.1",
