+++ src/Routes.js
@@ -2,18 +2,11 @@
 
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
-import Loadable from 'react-loadable';
 
 import TimeSheet from 'sections/TimeSheet';
 import Projects from 'sections/Projects';
 import Settings from 'sections/Settings';
-
-import Loading from './components/Loading';
-
-const Reports = Loadable({
-  loader: () => import('./sections/Reports'),
-  loading: Loading,
-});
+import Reports from 'sections/Reports';
 
 function Routes() {
   return (
